+++ lib/automatic_version_upgrader.dart
@@ -7,3 +7,4 @@ library automatic_version_upgrader;
 export 'src/command_runner.dart';
 export 'src/commands/commands.dart';
 export 'src/extensions/extensions.dart';
+export 'src/helpers/helpers.dart';